Extract GitHub transport protocol & shims
Introduce GitHubTransportProtocol and move the transport API surface out of GitHubURLSessionTransport into dedicated files. Add GitHubTransportShims (sharedGitHubTransport + backward-compat free-function shims) and move the public conformance methods into GitHubTransport+Conformance. Remove the protocol and shim implementations from GitHubURLSessionTransport (leaving the URLSession-backed struct), add TODO files documenting the two-part #1513 refactor, and bump .swiftlint.yml file_length warning slightly. This prepares the codebase for injecting/mockable transports as part of issue #1513.
